refactor: LadderGame을 LadderController로 이름 수정하고 디렉토리 변경
